% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapoolupdater_objects.R
\name{InstanceUpdate.error}
\alias{InstanceUpdate.error}
\title{InstanceUpdate.error Object}
\usage{
InstanceUpdate.error(InstanceUpdate.error.errors = NULL, errors = NULL)
}
\arguments{
\item{InstanceUpdate.error.errors}{The \link{InstanceUpdate.error.errors} object or list of objects}

\item{errors}{[Output Only] The array of errors encountered while processing this operation}
}
\value{
InstanceUpdate.error object
}
\description{
InstanceUpdate.error Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Errors that occurred during the instance update.
}
\seealso{
Other InstanceUpdate functions: \code{\link{InstanceUpdate.error.errors}},
  \code{\link{InstanceUpdate}}
}
