% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapoolupdater_objects.R
\name{InstanceUpdate}
\alias{InstanceUpdate}
\title{Google Compute Engine Instance Group Updater API Objects 
[Deprecated. Please use compute.instanceGroupManagers.update method. replicapoolupdater API will be disabled after December 30th, 2016] Updates groups of Compute Engine instances.}
\usage{
InstanceUpdate(InstanceUpdate.error = NULL,
  InstanceUpdate.error.errors = NULL, error = NULL, instance = NULL,
  status = NULL)
}
\arguments{
\item{InstanceUpdate.error}{The \link{InstanceUpdate.error} object or list of objects}

\item{InstanceUpdate.error.errors}{The \link{InstanceUpdate.error.errors} object or list of objects}

\item{error}{Errors that occurred during the instance update}

\item{instance}{Fully-qualified URL of the instance being updated}

\item{status}{Status of the instance update}
}
\value{
InstanceUpdate object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:09:30
filename: /Users/mark/dev/R/autoGoogleAPI/googlereplicapoolupdaterv1beta1.auto/R/replicapoolupdater_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
InstanceUpdate Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Update of a single instance.
}
\seealso{
Other InstanceUpdate functions: \code{\link{InstanceUpdate.error.errors}},
  \code{\link{InstanceUpdate.error}}
}
