% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapoolupdater_objects.R
\name{InstanceUpdateList}
\alias{InstanceUpdateList}
\title{InstanceUpdateList Object}
\usage{
InstanceUpdateList(items = NULL, nextPageToken = NULL, selfLink = NULL)
}
\arguments{
\item{items}{Collection of requested instance updates}

\item{nextPageToken}{A token used to continue a truncated list request}

\item{selfLink}{[Output Only] The fully qualified URL for the resource}
}
\value{
InstanceUpdateList object
}
\description{
InstanceUpdateList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response returned by ListInstanceUpdates method.
}

