% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapoolupdater_objects.R
\name{Operation}
\alias{Operation}
\title{Operation Object}
\usage{
Operation(Operation.error = NULL, Operation.error.errors = NULL,
  Operation.warnings = NULL, Operation.warnings.data = NULL,
  clientOperationId = NULL, creationTimestamp = NULL, endTime = NULL,
  error = NULL, httpErrorMessage = NULL, httpErrorStatusCode = NULL,
  id = NULL, insertTime = NULL, name = NULL, operationType = NULL,
  progress = NULL, region = NULL, selfLink = NULL, startTime = NULL,
  status = NULL, statusMessage = NULL, targetId = NULL,
  targetLink = NULL, user = NULL, warnings = NULL, zone = NULL)
}
\arguments{
\item{Operation.error}{The \link{Operation.error} object or list of objects}

\item{Operation.error.errors}{The \link{Operation.error.errors} object or list of objects}

\item{Operation.warnings}{The \link{Operation.warnings} object or list of objects}

\item{Operation.warnings.data}{The \link{Operation.warnings.data} object or list of objects}

\item{clientOperationId}{No description}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{endTime}{No description}

\item{error}{[Output Only] If errors occurred during processing of this operation, this field will be populated}

\item{httpErrorMessage}{No description}

\item{httpErrorStatusCode}{No description}

\item{id}{[Output Only] Unique identifier for the resource; defined by the server}

\item{insertTime}{[Output Only] The time that this operation was requested}

\item{name}{[Output Only] Name of the resource}

\item{operationType}{No description}

\item{progress}{No description}

\item{region}{[Output Only] URL of the region where the operation resides}

\item{selfLink}{[Output Only] The fully qualified URL for the resource}

\item{startTime}{[Output Only] The time that this operation was started by the server}

\item{status}{[Output Only] Status of the operation}

\item{statusMessage}{[Output Only] An optional textual description of the current status of the operation}

\item{targetId}{[Output Only] Unique target id which identifies a particular incarnation of the target}

\item{targetLink}{[Output Only] URL of the resource the operation is mutating}

\item{user}{No description}

\item{warnings}{No description}

\item{zone}{[Output Only] URL of the zone where the operation resides}
}
\value{
Operation object
}
\description{
Operation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An operation resource, used to manage asynchronous API requests.
}
\seealso{
Other Operation functions: \code{\link{Operation.error.errors}},
  \code{\link{Operation.error}},
  \code{\link{Operation.warnings.data}},
  \code{\link{Operation.warnings}}
}

