% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapoolupdater_objects.R
\name{Operation.warnings}
\alias{Operation.warnings}
\title{Operation.warnings Object}
\usage{
Operation.warnings(Operation.warnings.data = NULL)
}
\arguments{
\item{Operation.warnings.data}{The \link{Operation.warnings.data} object or list of objects}
}
\value{
Operation.warnings object
}
\description{
Operation.warnings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Operation functions: \code{\link{Operation.error.errors}},
  \code{\link{Operation.error}},
  \code{\link{Operation.warnings.data}},
  \code{\link{Operation}}
}

