% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapoolupdater_objects.R
\name{OperationList}
\alias{OperationList}
\title{OperationList Object}
\usage{
OperationList(id = NULL, items = NULL, nextPageToken = NULL,
  selfLink = NULL)
}
\arguments{
\item{id}{[Output Only] Unique identifier for the resource; defined by the server}

\item{items}{[Output Only] The Operation resources}

\item{nextPageToken}{[Output Only] A token used to continue a truncate}

\item{selfLink}{[Output Only] The fully qualified URL for the resource}
}
\value{
OperationList object
}
\description{
OperationList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains a list of Operation resources.
}

