% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapoolupdater_functions.R
\docType{package}
\name{replicapoolupdater_googleAuthR}
\alias{replicapoolupdater_googleAuthR}
\alias{replicapoolupdater_googleAuthR-package}
\title{Google Compute Engine Instance Group Updater API
Updates groups of Compute Engine instances.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:27:41
filename: /Users/mark/dev/R/autoGoogleAPI/googlereplicapoolupdaterv1beta1.auto/R/replicapoolupdater_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
\item https://www.googleapis.com/auth/replicapool
\item https://www.googleapis.com/auth/replicapool.readonly
}
}

