% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapoolupdater_objects.R
\name{RollingUpdate.error.errors}
\alias{RollingUpdate.error.errors}
\title{RollingUpdate.error.errors Object}
\usage{
RollingUpdate.error.errors()
}
\value{
RollingUpdate.error.errors object
}
\description{
RollingUpdate.error.errors Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other RollingUpdate functions: \code{\link{RollingUpdate.error}},
  \code{\link{RollingUpdate.policy}},
  \code{\link{RollingUpdate}},
  \code{\link{rollingUpdates.insert}}
}

