% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapoolupdater_objects.R
\name{RollingUpdate.error}
\alias{RollingUpdate.error}
\title{RollingUpdate.error Object}
\usage{
RollingUpdate.error(RollingUpdate.error.errors = NULL, errors = NULL)
}
\arguments{
\item{RollingUpdate.error.errors}{The \link{RollingUpdate.error.errors} object or list of objects}

\item{errors}{[Output Only] The array of errors encountered while processing this operation}
}
\value{
RollingUpdate.error object
}
\description{
RollingUpdate.error Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] Errors that occurred during the rolling update.
}
\seealso{
Other RollingUpdate functions: \code{\link{RollingUpdate.error.errors}},
  \code{\link{RollingUpdate.policy}},
  \code{\link{RollingUpdate}},
  \code{\link{rollingUpdates.insert}}
}
