% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapoolupdater_objects.R
\name{RollingUpdate.policy}
\alias{RollingUpdate.policy}
\title{RollingUpdate.policy Object}
\usage{
RollingUpdate.policy(autoPauseAfterInstances = NULL,
  instanceStartupTimeoutSec = NULL, maxNumConcurrentInstances = NULL,
  maxNumFailedInstances = NULL, minInstanceUpdateTimeSec = NULL)
}
\arguments{
\item{autoPauseAfterInstances}{Number of instances to update before the updater pauses the rolling update}

\item{instanceStartupTimeoutSec}{The maximum amount of time that the updater waits for a HEALTHY state after all of the update steps are complete}

\item{maxNumConcurrentInstances}{The maximum number of instances that can be updated simultaneously}

\item{maxNumFailedInstances}{The maximum number of instance updates that can fail before the group update is considered a failure}

\item{minInstanceUpdateTimeSec}{The minimum amount of time that the updater spends to update each instance}
}
\value{
RollingUpdate.policy object
}
\description{
RollingUpdate.policy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Parameters of the update process.
}
\seealso{
Other RollingUpdate functions: \code{\link{RollingUpdate.error.errors}},
  \code{\link{RollingUpdate.error}},
  \code{\link{RollingUpdate}},
  \code{\link{rollingUpdates.insert}}
}

