% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapoolupdater_objects.R
\name{RollingUpdateList}
\alias{RollingUpdateList}
\title{RollingUpdateList Object}
\usage{
RollingUpdateList(items = NULL, nextPageToken = NULL, selfLink = NULL)
}
\arguments{
\item{items}{Collection of requested updates}

\item{nextPageToken}{A token used to continue a truncated list request}

\item{selfLink}{[Output Only] The fully qualified URL for the resource}
}
\value{
RollingUpdateList object
}
\description{
RollingUpdateList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response returned by List method.
}

