% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapoolupdater_functions.R
\name{rollingUpdates.insert}
\alias{rollingUpdates.insert}
\title{Inserts and starts a new update.}
\usage{
rollingUpdates.insert(RollingUpdate, project, zone)
}
\arguments{
\item{RollingUpdate}{The \link{RollingUpdate} object to pass to this method}

\item{project}{The Google Developers Console project name}

\item{zone}{The name of the zone in which the update's target resides}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/replicapool
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/replicapool)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/compute/docs/instance-groups/manager/#applying_rolling_updates_using_the_updater_service}{Google Documentation}

Other RollingUpdate functions: \code{\link{RollingUpdate.error.errors}},
  \code{\link{RollingUpdate.error}},
  \code{\link{RollingUpdate.policy}},
  \code{\link{RollingUpdate}}
}
