% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapoolupdater_functions.R
\name{rollingUpdates.pause}
\alias{rollingUpdates.pause}
\title{Pauses the update in state from ROLLING_FORWARD or ROLLING_BACK. Has no effect if invoked when the state of the update is PAUSED.}
\usage{
rollingUpdates.pause(project, zone, rollingUpdate)
}
\arguments{
\item{project}{The Google Developers Console project name}

\item{zone}{The name of the zone in which the update's target resides}

\item{rollingUpdate}{The name of the update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/replicapool
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/replicapool)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/compute/docs/instance-groups/manager/#applying_rolling_updates_using_the_updater_service}{Google Documentation}
}
